[SERVER] refactor: 아이템 버리기 시 버려지는 아이템 개수 1개로 고정되게 수정

+++ server/realtime/src/main/java/com/explorer/realtime/gamedatahandling/inventory/event/DropItemInInventory.java
@@ -4,6 +4,8 @@ import com.explorer.realtime.gamedatahandling.component.personal.inventoryInfo.d
 import com.explorer.realtime.gamedatahandling.component.personal.inventoryInfo.repository.InventoryRepository;
 import com.explorer.realtime.gamedatahandling.farming.repository.MapInfoRepository;
 import com.explorer.realtime.gamedatahandling.inventory.dto.ItemInfo;
+import com.explorer.realtime.gamedatahandling.inventory.exception.InventoryErrorCode;
+import com.explorer.realtime.gamedatahandling.inventory.exception.InventoryException;
 import com.explorer.realtime.global.common.dto.Message;
 import com.explorer.realtime.global.common.enums.CastingType;
 import com.explorer.realtime.global.component.broadcasting.Broadcasting;
@@ -35,34 +37,47 @@ public class DropItemInInventory {
         int mapId = json.getInt("mapId");
         Long userId = json.getLong("userId");
         int inventoryIdx = json.getInt("inventoryIdx");
-        int itemCnt = json.getInt("itemCnt");
+
         String position = json.getString("position");
-        log.info("[process] channelId : {}, mapId : {} , userId : {}, inventoryIdx : {}, itemCnt : {}, position : {}", channelId, mapId, userId, inventoryIdx, itemCnt, position);
+        log.info("[process] channelId : {}, mapId : {} , userId : {}, inventoryIdx : {}, position : {}", channelId, mapId, userId, inventoryIdx, position);
 
         return inventoryRepository.findByInventoryIdx(channelId, userId, inventoryIdx)
                 .flatMap(inventory -> {
+                    if (String.valueOf(inventory).isEmpty()) {
+                        return Mono.error(new InventoryException(InventoryErrorCode.EMPTY_INVENTORY));
+                    }
+
                     InventoryInfo inventoryInfo = InventoryInfo.ofString(inventoryIdx, String.valueOf(inventory));
                     String itemCategory = inventoryInfo.getItemCategory();
                     int itemId = inventoryInfo.getItemId();
                     log.info("[process] inventoryInfo : {}", inventoryInfo);
 
-                    int result = inventoryInfo.getItemCnt() - itemCnt;
+                    int result = inventoryInfo.getItemCnt() - 1;
                     if (result <= 0) {
                         return inventoryRepository.deleteByInventoryIdx(channelId, userId, inventoryIdx)
                                 .then(Mono.just(InventoryInfo.ofString(inventoryIdx, "")))
                                 .flatMap(emptyInventoryInfo -> {
-                                    return mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId, itemCnt)
+                                    return mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId)
                                             .then(unicasting(channelId, userId, emptyInventoryInfo))
-                                            .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId, itemCnt)));
+                                            .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId)));
                                 });
                     } else {
                         inventoryInfo.setItemCnt(result);
                         return inventoryRepository.save(channelId, userId, inventoryInfo)
-                                .then(mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId, itemCnt))
+                                .then(mapInfoRepository.save(channelId, mapId, position, itemCategory, itemId))
                                 .then(unicasting(channelId, userId, inventoryInfo))
-                                .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId, itemCnt)));
+                                .then(broadcasting(channelId, position, ItemInfo.of(itemCategory, itemId)));
                     }
-                });
+                })
+                .onErrorResume(InventoryException.class, error -> {
+                    log.info("[process] errorCode : {}, errorMessage : {}", error.getErrorCode(), error.getMessage());
+                    unicasting.unicasting(
+                            channelId,
+                            userId,
+                            MessageConverter.convert(Message.fail(eventName, CastingType.UNICASTING, String.valueOf(error.getErrorCode()), error.getMessage()))
+                    ).subscribe();
+                    return Mono.empty();
+                }).then();
     }
 
     public Mono<Void> unicasting(String channelId, Long userId, InventoryInfo inventoryInfo) {
